Handle partially-read lines in stream mode
When input is sent via stdin, it is possible that read() reads the input
faster than it is written, which can result in the read loop terminating
with a partially-read line (meaning no newline is present at the end of
the input). This part of the input should be preserved until more data
can be read from stdin, so incomplete trailing lines need to be
preserved across calls to stream::on_stream_data.